Introduce Tabs Control
Replicates the appearance of the standard W95 tabs with clickable labels
that show and hide the associated content regions. There are many
different ways to implement tabs in React. This approach is mostly aimed
at ensuring no state management stuff is required for users of the
library and ensuring that the tabs look similar to the existing
components at the top level (though this is complicated by the need to
have `Tab` components nested inside `Tabs`, a constraint that might
not be as simple as other parts of the library).

Currently doesn’t support custom props for events (`onEnterTab`, `onLeaveTab`,
etc) and doesn’t have an explicit API for focusing/switching to a tab
programatically. All this would be possible to add, just want to start
with something simple that might be useful.